pyproject.toml
[project]
name = "chatgpt-telegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.4",
    "requests>=2.32.5",
]
